chore(cli): add command size advisory
